LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/fetch/api/policies/csp-blocked-expected.txt


PASS Fetch is blocked by CSP, got a TypeError
